% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0_createObject.R
\docType{class}
\name{CellInter-class}
\alias{CellInter-class}
\alias{CellInter}
\title{The CellInter Class
The CellInter object is the center of each cell-cell communication analysis with the scRNA-seq.}
\description{
The CellInter Class
The CellInter object is the center of each cell-cell communication analysis with the scRNA-seq.
}
\section{Slots}{

\describe{
\item{\code{data}}{The normalized expression matrix, mean matrix, gsea result, lr score matrix, and lr score matrix(log-scale)}

\item{\code{meta.data}}{cell type, barcode, nFeature and count of each cell}

\item{\code{reductions}}{data.frame of L-R-TF with specific cellA-cellB, later for sankey plot}

\item{\code{project}}{the name of this analysis}

\item{\code{Org}}{the species of this scRNA-seq}

\item{\code{version}}{the version of this R package}
}}

