% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2rpm.R
\name{counts2normalized_10X}
\alias{counts2normalized_10X}
\title{transform count to CPM}
\usage{
counts2normalized_10X(data, toType, scale.factor = 10^6)
}
\arguments{
\item{data}{a dataframe with row of gene and column of sample}

\item{toType}{whether RPKM or TPM of result you wanted, eg "CPM"}

\item{scale.factor}{set the scale factor, default "10^6"}
}
\value{
the CPM of \code{data}
}
\description{
transform count to CPM
}
