% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2rpm.R
\name{counts2normalized_smartseq2}
\alias{counts2normalized_smartseq2}
\title{transform count to RPKM or TPM within Hsapiens and Mmusculus}
\usage{
counts2normalized_smartseq2(data, Org, toType, scale.factor = 10^6)
}
\arguments{
\item{data}{a dataframe with row of gene and column of sample}

\item{Org}{choose the species source of gene, eg "Homo sapiens", "Mus musculus"}

\item{toType}{whether RPKM or TPM of result you wanted, eg "RPKM", "TPM"}

\item{scale.factor}{set the scale factor, default "10^6"}
}
\value{
the RPKM or TPM of \code{data}
}
\description{
transform count to RPKM or TPM within Hsapiens and Mmusculus
}
