% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0_createObject.R
\name{CreateNichConObject}
\alias{CreateNichConObject}
\title{create a Cellwave objects}
\usage{
CreateNichConObject(
  data,
  min.feature = 3,
  names.field = 1,
  names.delim = "_",
  project = "Microenvironment",
  source = "UMI",
  scale.factor = 10^6,
  Org = "Homo sapiens"
)
}
\arguments{
\item{data}{a dataframe with row of gene and column of sample}

\item{min.feature}{Include cells where at least this many features are detected}

\item{names.field}{BARCODE_CELLTYPE in the input matrix, set names.field to 2 to set the initial identities to CELLTYPE.}

\item{names.delim}{For the initial identity class for each cell, choose this delimiter from the cell's column name. E.g. If your cells are named as BARCODE_CELLTYPE, set this to "_" to separate the cell name into its component parts for picking the relevant field.}

\item{project}{Project name for the this object}

\item{source}{the type of expression dataframe, eg "UMI", "fullLength", "TPM", or "CPM"}

\item{scale.factor}{set the scale factor, default "10^6"}

\item{Org}{choose the species source of gene, eg "Homo sapiens", "Mus musculus"}
}
\value{
the value of \code{cellwave object}
}
\description{
create a Cellwave objects
}
