% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelatedGene.R
\name{getCorrelatedGene}
\alias{getCorrelatedGene}
\title{get relation among specific TF}
\usage{
getCorrelatedGene(
  data,
  cell_type = "",
  tf,
  target_list,
  pValue = 0.05,
  corValue = 0,
  topGene = 0.1
)
}
\arguments{
\item{data}{a dataframe with row of gene and column of sample}

\item{cell_type}{specific cell type}

\item{tf}{choose one specific TF}

\item{target_list}{choose target genes corresponding TF}

\item{pValue}{set the pValue of spearman}

\item{corValue}{set the corValue of spearman}

\item{topGene}{use topTargetCor of candidate genes which has firlter by above parameters, default is 1, means 100\%}
}
\value{
result target genes in this regulon
}
\description{
get relation among specific TF
}
