% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2Bubble.R
\name{getForBubble}
\alias{getForBubble}
\title{melt enrich dataframe for bubble graph}
\usage{
getForBubble(pathway.hyper.list = pathway.hyper.list)
}
\arguments{
\item{pathway.hyper.list}{a list of pathway enrichment pvalue and NES}
}
\value{
the result for pbubble graph
}
\description{
melt enrich dataframe for bubble graph
}
