% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{getGSEA}
\alias{getGSEA}
\title{calculate enrich for each celltype with all regulons}
\usage{
getGSEA(term_gene_list, FC_OF_CELL, minGSSize = 5, maxGSSize = 500)
}
\arguments{
\item{term_gene_list}{all regulons of genes denotes geneSet}

\item{FC_OF_CELL}{the cell type which you want to calculate the foldchange}

\item{minGSSize}{set the min size of each geneSet}

\item{maxGSSize}{set the max size of each geneSet}
}
\value{
the enrich result of each celltype with all regulons
}
\description{
calculate enrich for each celltype with all regulons
}
