% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGSEAplot.R
\name{getGSEAplot}
\alias{getGSEAplot}
\title{plot enrichment graph}
\usage{
getGSEAplot(
  gsea.list,
  myCelltype,
  fc.list,
  geneSetID,
  selectedGeneID = NULL,
  mycol = NULL
)
}
\arguments{
\item{gsea.list}{a list of enrichment result from cellwave}

\item{myCelltype}{the cell type of receiver cell}

\item{fc.list}{foldchange list in the cellwave object}

\item{geneSetID}{the character of TF symbol, only significant activated can be inspected}

\item{selectedGeneID}{default is NULL, label the position of specific gene in FC flow.}

\item{mycol}{the color of each TF. the length is consistent with geneSetID}
}
\description{
plot enrichment graph
}
