% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignificantPathway.R
\name{getHyperPathway}
\alias{getHyperPathway}
\title{enrich communication relation on the pathway}
\usage{
getHyperPathway(data, cella_cellb, IS_core = TRUE, Org = "Homo sapiens")
}
\arguments{
\item{data}{a dataframe of communication score with row LR and column cellA-cellB}

\item{cella_cellb}{explore the LR between sender cellA and receiver cellB, eg: "A-B"}

\item{IS_core}{logical variable ,whether use reference LR data or include extended datasets}

\item{Org}{choose the species source of gene, only "Homo sapiens" in this version.}
}
\value{
the dataframe with column: Pvalue, Jaccard, NES and pathway
}
\description{
enrich communication relation on the pathway
}
