% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR2TF.R
\name{LR2TF}
\alias{LR2TF}
\title{get score of Triple relation corresponding specific cellA-cellB}
\usage{
LR2TF(
  object,
  sender_cell,
  recevier_cell,
  slot = "expr_l_r_log2_scale",
  org = "Homo sapiens"
)
}
\arguments{
\item{object}{a Cellwave objects}

\item{sender_cell}{the cell type of sender cell}

\item{recevier_cell}{the cell type of recevier cell}

\item{slot}{plot the graph with the data of specific slot}

\item{org}{choose the species source of gene, eg "Homo sapiens", "Mus musculus"}
}
\description{
get score of Triple relation corresponding specific cellA-cellB
}
