% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBubble.R
\name{plotBubble}
\alias{plotBubble}
\title{plot bubble graph}
\usage{
plotBubble(dat)
}
\arguments{
\item{dat}{a dataframe of melted communication score with columns: pathway, cc, pvalue, NES, logPvalue}
}
\description{
plot bubble graph
}
