% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfRidgeplot.R
\name{ridgeplot.DIY}
\alias{ridgeplot.DIY}
\title{plot ridge graph}
\usage{
ridgeplot.DIY(
  x,
  showCategory = targetTFindex,
  fill = "p.adjust",
  core_enrichment = TRUE,
  orderBy = "NES",
  decreasing = FALSE
)
}
\arguments{
\item{x}{gseaResult object}

\item{showCategory}{number of categories for plotting}

\item{fill}{one of "pvalue", "p.adjust", "qvalue"}

\item{core_enrichment}{whether only using core_enriched genes}

\item{orderBy}{The order of the Y-axis, default is NES, other colnames is ok, eg: "ID", "Description", "setSize", "enrichmentScore", "p.adjust".}

\item{decreasing}{logical. Should the orderBy order be increasing or decreasing?}
}
\description{
plot ridge graph
}
