% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sankey_graph}
\alias{sankey_graph}
\title{plot circle graph with communication profile}
\usage{
sankey_graph(
  df,
  axes,
  mycol = NULL,
  nudge_x = NULL,
  font.size = 5,
  boder.col = "black",
  isGrandSon = FALSE,
  set_alpha = 1
)
}
\arguments{
\item{df}{a dataframe with five or four columns depending on the parameter isGrandSon}

\item{axes}{if triple realtion of sankey, set axes=1:3, otherwise bi is 1:2. default 1:3}

\item{mycol}{a vector of character, denotes the color of each node}

\item{nudge_x}{a vector of numeric, denotes the horizontal position of each node label}

\item{font.size}{the font size of node label}

\item{boder.col}{the color of node border}

\item{isGrandSon}{If FALSE, the third axe inherits the first two axes and only consider about relation instead of score. otherwise every axe only inherit the first one axes ggtitle and consider about score.}

\item{set_alpha}{set the alpha of color in the node, a numeric bwtween 0-1.}
}
\description{
plot circle graph with communication profile
}
