% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shellyFunction.R
\name{trans2tripleScore}
\alias{trans2tripleScore}
\title{get score of Triple relation corresponding specific cellA-cellB}
\usage{
trans2tripleScore(df)
}
\arguments{
\item{df}{a dataframe consist of five columns: "Ligand", "Receptor", "TF", "weight1", "weight2"}
}
\value{
a dataframe consist of four columns: "Ligand", "Receptor", "TF", "weight"
}
\description{
get score of Triple relation corresponding specific cellA-cellB
}
