% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewInterCircos.R
\name{ViewInterCircos}
\alias{ViewInterCircos}
\title{plot circle graph with communication profile}
\usage{
ViewInterCircos(
  object,
  font = 2,
  cellColor,
  lrColor = NULL,
  order.vector = NULL,
  trackhight1 = 0.05,
  linkcolor.from.sender = TRUE,
  linkcolor = NULL,
  arr.type = "big.arrow",
  arr.length = 0.04,
  DIY = TRUE,
  gap.degree = NULL,
  trackhight2 = 0.032,
  track.margin2 = c(0.01, 0.12),
  slot = "expr_l_r_log2_scale"
)
}
\arguments{
\item{object}{a Cellwave objects}

\item{font}{the size of font}

\item{cellColor}{a color dataframe, rownames is cell type, value is color}

\item{lrColor}{a color vector denotes the color of ligand and receptor, containing two elements, default is c('#D92E27', "#35C6F4")}

\item{order.vector}{default is null, a celltype vector with the order you want in the circle graph}

\item{trackhight1}{Height of the outer track}

\item{linkcolor.from.sender}{logical value, whether the color of line correspond with color of sender cell}

\item{linkcolor}{one color you want link to be, only if parameter linkcolor.from.sender=FALSE}

\item{arr.type}{Type of the arrows, default value is big.arrow There is an additional option triangle}

\item{arr.length}{Length of the arrows, measured in 'cm'. If arr.type is set to big.arrow, the value is percent to the radius of the unit circle.}

\item{DIY}{logical value, if TRUE, the parameter object should be a dataframe, and set slot="expr_l_r_log2_scale". otherwise object should be a Cellwave objects.}

\item{gap.degree}{between two neighbour sectors. It can be a single value or a vector. If it is a vector, the first value corresponds to the gap after the first sector.}

\item{trackhight2}{Height of the inner track}

\item{track.margin2}{affect current track}

\item{slot}{plot the graph with the data of specific slot}
}
\description{
plot circle graph with communication profile
}
