% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPheatmap.R
\name{viewPheatmap}
\alias{viewPheatmap}
\title{plot pheatmap graph with communication profile}
\usage{
viewPheatmap(
  object,
  slot = "expr_l_r_log2_scale",
  show_rownames = T,
  show_colnames = T,
  treeheight_row = 0,
  treeheight_col = 50,
  cluster_rows = T,
  cluster_cols = F,
  fontsize = 12,
  angle_col = "45",
  color = NULL,
  main = "score"
)
}
\arguments{
\item{object}{a Cellwave objects}

\item{slot}{plot the graph with the data of specific slot}

\item{show_rownames}{boolean specifying if column names are be shown.}

\item{show_colnames}{boolean specifying if column names are be shown.}

\item{treeheight_row}{the height of a tree for rows, if these are clustered. Default value 0 points.}

\item{treeheight_col}{the height of a tree for columns, if these are clustered. Default value 50 points.}

\item{cluster_rows}{boolean values determining if rows should be clustered or hclust object,}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object.}

\item{fontsize}{base fontsize for the plot}

\item{angle_col}{angle of the column labels, right now one can choose only from few predefined options (0, 45, 90, 270 and 315)}

\item{color}{vector of colors used in heatmap.}

\item{main}{the title of the plot, default is "score".}
}
\description{
plot pheatmap graph with communication profile
}
